Refactor data fetching and adjust order of tabs
